fix: memory cache
