+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/gasoid/merge-bot/cache"
 	"github.com/gasoid/merge-bot/config"
 	_ "github.com/gasoid/merge-bot/handlers/gitlab"
 	"github.com/gasoid/merge-bot/logger"
@@ -16,6 +17,10 @@ func main() {
 	}
 
 	logger.New()
+
+	if err := cache.Init(); err != nil {
+		logger.Error("cache can't be initialized", "error", err)
+	}
 
 	startMetricsEndpoint()
 	start()
